♻️ : 유틸리티 함수 분리
